Refactor: Moved app.py to __init__.py and moved into app folder.
