+++ app/__init__[OBSOLETE].py
@@ -1,0 +1,44 @@
+import os
+
+from flask import Flask, render_template, url_for
+
+
+def create_app(test_config=None):
+    # Create and configure the app
+    app = Flask(__name__, instance_relative_config=True)
+    app.config.from_mapping(
+        SECRET_KEY="MUSTCHANGE",
+        DATABASE=os.path.join(app.instance_path, "data-dev.sqlite"),
+    )
+
+    if test_config is None:
+        # load the instance config, if it exists, when not testing
+        app.config.from_pyfile("config.py", silent=True)
+    else:
+        # load the test config if passed in
+        app.config.from_mapping(test_config)
+
+    # ensure the instance folder exists
+    try:
+        os.makedirs(app.instance_path)
+    except OSError:
+        pass
+
+    # a simple page that says hello
+    @app.route("/hello")
+    def hello():
+        return "Hello, World!"
+
+    @app.route("/")
+    def index():
+        return render_template("index.html")
+
+    from . import db
+
+    db.init_app(app)
+
+    from . import auth
+
+    app.register_blueprint(auth.bp)
+
+    return app
