+++ app/templates/new_folder.html
