+++ app/templates[OBSOLETE]/index.html
@@ -1,0 +1,1 @@
+{% extends 'base.html' %}
